ShowCreditsLoader: extract TV show id finder to TmdbTools2.
