+++ app/src/main/java/com/battlelancer/seriesguide/ui/people/ShowCreditsLoader.kt
@@ -2,15 +2,11 @@ package com.battlelancer.seriesguide.ui.people
 
 import android.content.Context
 import com.battlelancer.seriesguide.SgApp
+import com.battlelancer.seriesguide.tmdbapi.TmdbTools2
 import com.battlelancer.seriesguide.util.Errors
 import com.uwetrottmann.androidutils.GenericSimpleLoader
 import com.uwetrottmann.tmdb2.entities.Credits
-import com.uwetrottmann.tmdb2.enumerations.ExternalSource
-import com.uwetrottmann.tmdb2.services.FindService
-import com.uwetrottmann.tmdb2.services.TvService
-import dagger.Lazy
 import timber.log.Timber
-import javax.inject.Inject
 
 /**
  * Create a show credit [android.support.v4.content.Loader]. Supports show ids from TVDb
@@ -18,30 +14,32 @@ import javax.inject.Inject
  *
  * @param findTmdbId If true, the loader assumes the passed id is from TVDb id and will try to
  */
-class ShowCreditsLoader(context: Context, private var showId: Int, private val findTmdbId: Boolean)
-    : GenericSimpleLoader<Credits?>(context) {
-
-    @Inject
-    lateinit var findService: Lazy<FindService>
-    @Inject
-    lateinit var tvService: Lazy<TvService>
-
-    init {
-        SgApp.getServicesComponent(context).inject(this)
-    }
+class ShowCreditsLoader(
+    context: Context,
+    private var showId: Int,
+    private val findTmdbId: Boolean
+) : GenericSimpleLoader<Credits?>(context) {
 
     override fun loadInBackground(): Credits? {
-        if (findTmdbId && !findShowTmdbId()) {
-            return null // failed to find the show on TMDb
+        if (findTmdbId) {
+            val tmdbId = TmdbTools2(context).findShowTmdbId(showId)
+            showId = if (tmdbId != null && tmdbId > 0) {
+                tmdbId
+            } else {
+                Timber.d("Downloading show credits failed: show not on TMDb")
+                return null // Failed to find the show on TMDb.
+            }
         }
 
-        if (showId < 0) {
-            return null // do not have a valid id, abort
+        if (showId <= 0) {
+            return null // Do not have a valid id, abort.
         }
 
         // get credits for that show
         try {
-            val response = tvService.get().credits(showId, null).execute()
+            val response = SgApp.getServicesComponent(context).tmdb().tvService()
+                .credits(showId, null)
+                .execute()
             if (response.isSuccessful) {
                 return response.body()
             } else {
@@ -54,28 +52,4 @@ class ShowCreditsLoader(context: Context, private var showId: Int, private val f
         return null
     }
 
-    private fun findShowTmdbId(): Boolean {
-        try {
-            val response = findService.get()
-                    .find(showId.toString(), ExternalSource.TVDB_ID, null)
-                    .execute()
-            if (response.isSuccessful) {
-                val tvResults = response.body()?.tv_results
-                if (!tvResults.isNullOrEmpty()) {
-                    val showId = tvResults[0].id
-                    showId?.let {
-                        this.showId = showId
-                        return true // found it!
-                    }
-                }
-                Timber.d("Downloading show credits failed: show not on TMDb")
-            } else {
-                Errors.logAndReport("find tvdb show", response)
-            }
-        } catch (e: Exception) {
-            Errors.logAndReport("find tvdb show", e)
-        }
-
-        return false
-    }
 }
